require 'sendupdates' file in calendars dir to send updates.

+++ readrooms.php
@@ -1,6 +1,10 @@
 #!/usr/bin/env php
 <?php 
 
+if ($GLOBALS['argv'][1] === "check" && file_exists("/calendars/GLRF_ALL.ics")) {
+	exit(1);
+}
+
 chdir("/opt/roardaemon");
 
 system("wget -O rooms2.xml --user tomh --password 25LiveSFSPublisher --no-check-certificate 'https://webservices.collegenet.com/r25ws/wrd/uwm/run/spaces.xml?ML_FLS=R&name=GLRF&scope=list'");
